- Started a switch to uber shaders (options keyword) because subroutines are clunky and no one uses them. - Removed techniques and passes because they are going to be redundant - Resources are managed in their own buffer - much friendlier to the next wave of graphics APIs and bindless is still usable. - Fixed textures binding partially. I should probably add resource layouts for parsing baked tables.

+++ include/tempest/shader/shader-convert-common.hh
@@ -38,7 +38,7 @@ class Variable;
 class BufferDescription;
 class EffectDescription;
 
-size_t ConvertStructBuffer(const Variable* var, Shader::EffectDescription* fx_desc);
+uint32 ConvertStructBuffer(const Variable* var, Shader::EffectDescription* fx_desc);
 void ConvertBuffer(const Buffer* buffer, Shader::EffectDescription* fx_desc);
 }
 }
